fix shorten-64-to-32 warning in crc32c

+++ util/crc32c.cc
@@ -319,9 +319,9 @@ static inline void Fast_CRC32(uint64_t* l, uint8_t const **p) {
   *l = _mm_crc32_u64(*l, LE_LOAD64(*p));
   *p += 8;
 #else
-  *l = _mm_crc32_u32(*l, LE_LOAD32(*p));
+  *l = _mm_crc32_u32((unsigned int)*l, LE_LOAD32(*p));
   *p += 4;
-  *l = _mm_crc32_u32(*l, LE_LOAD32(*p));
+  *l = _mm_crc32_u32((unsigned int)*l, LE_LOAD32(*p));
   *p += 4;
 #endif
 #else
